Move over MCL pairing code to interface implementation
